services/analyst/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "level-analyst"
version = "0.1.0"
description = "AI Analyst module for Level Parking"
authors = [
    {name = "Level Parking", email = "engineering@lvlparking.com"},
]
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "gunicorn>=21.2.0",
    "pydantic>=2.4.0",
    "pydantic-settings>=2.0.0",
    "psycopg[binary]>=3.1.0",
    "asyncpg>=0.28.0",
    "apscheduler>=3.10.4",
    "sqlalchemy>=2.0.0",
    "alembic>=1.12.0",
    "openai>=1.0.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.6",
    "httpx>=0.25.0",
    "pandas>=2.1.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
    "pgvector>=0.2.0",
    "pytz>=2023.3",
    "opentelemetry-sdk>=1.27.0",
    "opentelemetry-exporter-otlp>=1.27.0",
    "opentelemetry-instrumentation-fastapi>=0.48b0",
    "prometheus_client>=0.20.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.25.0",
    "pytest-mock>=3.11.0",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["analyst"]

[tool.hatch.build.targets.sdist]
include = [
    "analyst",
    "pyproject.toml",
]
